G3 2023 v5 issue#13821

+++ DuggaSys/fileed.js
@@ -346,11 +346,14 @@ function renderCell(col, celldata, cellid) {
     if (col == "trashcan" || col == "filename" || col == "filesize" || col == "editor") {
         obj = JSON.parse(celldata);
     }
-
     if (col == "trashcan") {
         if (obj.showtrashcan) {
-            str = "<span class='iconBox'><img alt='delete file icon' tabindex='0' id='dorf' title='Delete file' class='trashcanIcon trashcanTab' src='../Shared/icons/Trashcan.svg' ";
-            str += " onclick='deleteFile(\"" + obj.fileid + "\",\"" + obj.filename + "\",\"" + obj.filekind + "\");' ></span>";
+            if(obj.filePath.includes("Github")) {
+                str = "<span class='iconBox'><img alt='github icon' tabindex='0' id='dorf' title='Github file' style='width:16px' src='../Shared/icons/githubLink-icon.png' ";
+            } else {
+                str = "<span class='iconBox'><img alt='delete file icon' tabindex='0' id='dorf' title='Delete file' class='trashcanIcon trashcanTab' src='../Shared/icons/Trashcan.svg' ";
+                str += " onclick='deleteFile(\"" + obj.fileid + "\",\"" + obj.filename + "\",\"" + obj.filekind + "\");' ></span>";
+            }
         }
     } else if (col == "filename") {
         if (obj.kind == "Link") {
@@ -367,15 +370,17 @@ function renderCell(col, celldata, cellid) {
     } else if (col == "extension" || col == "uploaddate") {
         str += "<span>" + celldata + "</span>";
     } else if (col == "editor") {
-        if(obj.showeditor){
-        if (obj.extension == "md" || obj.extension == "txt" || obj.extension == "html") {
-            str = "<span class='iconBox' ><img alt='edit file icon' tabindex='0' id='dorf'  title='Edit file' class='markdownIcon markdownIconeditFile' src='../Shared/icons/markdownPen.svg' ";
-            str += "onclick='loadPreview(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
-        } else if (obj.extension == "js"  || obj.extension == "css" || obj.extension == "php") {
-            str = "<span class='iconBox'><img alt='edit file icon' id='dorf'  title='Edit file'  class='markdownIcon' src='../Shared/icons/markdownPen.svg' ";
-            str += "onclick='loadFile(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
+        if (obj.showeditor) {
+            if (!obj.filePath.includes("Github")) {
+                if (obj.extension == "md" || obj.extension == "txt" || obj.extension == "html") {
+                    str = "<span class='iconBox' ><img alt='edit file icon' tabindex='0' id='dorf'  title='Edit file' class='markdownIcon markdownIconeditFile' src='../Shared/icons/markdownPen.svg' ";
+                    str += "onclick='loadPreview(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
+                } else if (obj.extension == "js" || obj.extension == "css" || obj.extension == "php") {
+                    str = "<span class='iconBox'><img alt='edit file icon' id='dorf'  title='Edit file'  class='markdownIcon' src='../Shared/icons/markdownPen.svg' ";
+                    str += "onclick='loadFile(\"" + obj.filePath + "\", \"" + obj.filename + "\", " + obj.kind + ")'></span>";
+                }
+            }
         }
-    }
     } else if (col == "kind") {
         str += "<span>" + convertFileKind(celldata) + "</span>";
     } else if (col == "type") {
